Functions

Index

PowerModelsMCDC.build_mcdcopf

PowerModelsMCDC.parse_file

PowerModelsMCDC.solve_mcdcopf

Working with input data

PowerModelsMCDC.parse_file — Function

parse_file(file; <keyword arguments>)

Parse a Matpower .m file into a PowerModelsMCDC data structure.

Keyword arguments, if any, are forwarded to PowerModels.parse_file.

Problem specifications

PowerModelsMCDC.solve_mcdcopf — Function

solve_mcdcopf(file, model_type, optimizer; <keyword arguments>)
solve_mcdcopf(data, model_type, optimizer; <keyword arguments>)

Build and solve the OPF problem over a hybrid AC/DC network, using a multi-conductor model for the DC part.

Input can be a Matpower file or a data dictionary. The OPF problem being built is the one defined in build_mcdcopf. Keyword arguments, if any, are forwarded to PowerModels.solve_model.

PowerModelsMCDC.build_mcdcopf — Function

build_mcdcopf(pm::PowerModels.AbstractPowerModel)

Build the OPF problem over a hybrid AC/DC network, using a multi-conductor model for the DC part.

The objective is the minimization of generation cost.